shuo/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "shuo"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "websockets>=12.0",
    "twilio>=9.0.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "pydantic-ai-slim[groq]>=0.2.0",
    "deepgram-sdk>=3.0.0",
    "httpx>=0.27.0",
    "click>=8.0.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
voice-agent = "shuo.cli:main"
